[TileK] Fixes for gcc 4.2.4 and 4.8.1 (template issues: 4.2.4 was confused by collectReferencedSymbols and 4.8.1 is pedantic). Replaced all occurence of unsigned and unsigned long by size_t.

+++ src/backend/KLT/lib/tilek/tilek.cpp
@@ -18,14 +18,9 @@ typedef ::MDCG::KLT::Runtime<Hlang, Klang> Runtime; // Runtime Description
 
 namespace KLT {
 
-template <>
-size_t ::KLT::LoopTrees<Annotation>::id_cnt = 0;
-
-template <>
-unsigned long ::KLT::Kernel<Annotation, Runtime>::id_cnt = 0;
-
-template <>
-unsigned long ::KLT::Kernel<Annotation, Runtime>::kernel_desc_t::id_cnt = 0;
+template <> size_t ::KLT::LoopTrees<Annotation>::id_cnt = 0;
+template <> size_t ::KLT::Kernel<Annotation, Runtime>::id_cnt = 0;
+template <> size_t ::KLT::Kernel<Annotation, Runtime>::kernel_desc_t::id_cnt = 0;
 
 template <>
 SgFunctionParameterList * Kernel<Annotation, Runtime>::createParameterList() const {
